Check if the response generates an error
If the call the read the response from the server generates an error, we should return from the function instead of looping indefinitely.

I had a situation where the NUT master's ACL wasn't configured correctly. The master would close the socket connection whenever the client tried to connect. This would cause the client to be stuck in a busy loop.